Drawing wedge for MinMaxCriterion
function DrawingWedge(int number), int number - the number of criterion
in CriterionFactory

+++ TurchinaTPRLab/Core/Service/Criterions/Randomized criterions/RandBayesCriterion.cs
@@ -60,7 +60,7 @@ namespace TurchinaTPRLab.Core.Service.Criterions.Randomized_criterions
                     }
                     else
                     {
-                        tempArray[i] = 1 / countOfMinPoints;
+                        tempArray[i] = (double) 1 / countOfMinPoints;
                     }
                 }
                 return new Solution(tempArray);
